added force document | quote it fix attempt | black spaces

+++ userge/plugins/jutsu/quoteIT_for_ux.py
@@ -1,9 +1,10 @@
-
 import asyncio
 import json
 import os
+import re
 import time
 
+from pyrogram import filters
 from userge import userge, Message
 
 
@@ -21,26 +22,32 @@ from userge import userge, Message
 async def make_tweet(message: Message):
     try:
         await userge.get_chat_member(-1001331162912, message.from_user.id)
-    except:
-        return await message.edit("First join **@UX_xplugin_support** and get approved by Kakashi.")
+    except BaseException:
+        return await message.edit(
+            "First join **@UX_xplugin_support** and get approved by Kakashi."
+        )
     reply_ = message.replied
     if not reply_:
         return await message.edit("`Reply to message...`", del_in=5)
     name_ = reply_.from_user.first_name
     username_ = "@" + reply_.from_user.username
     pfp_ = reply_.from_user.photo.big_file_id or None
-    text_ = reply_.text if reply_.text and "-f" not in message.flags else message.filtered_input_str
+    text_ = (
+        reply_.text
+        if reply_.text and "-f" not in message.flags
+        else message.filtered_input_str
+    )
     bg_ = (26, 43, 60) if "-b" not in message.flags else (0, 0, 0)
     if not text_:
         return await message.edit("`Text not found...`", del_in=5)
     await message.edit("`Making tweet...`")
     bot_ = "QuoteIT_thebot"
     form_ = {
-        'cmd': 'TWEET_IT',
-        'name': name_,
-        'username': username_,
-        'text': text_,
-        'background': bg_,
+        "cmd": "TWEET_IT",
+        "name": name_,
+        "username": username_,
+        "text": text_,
+        "background": bg_,
     }
     json_ = json.dumps(form_, indent=4)
     if pfp_:
@@ -48,25 +55,24 @@ async def make_tweet(message: Message):
         msg_ = await userge.send_photo(bot_, down_, caption=json_)
         os.remove(down_)
     else:
-        msg_ = await userge.send_message(bot_, json_)
-    await asyncio.sleep(7.5)
-    start_time = time.time()
-    while True:
-        try:
-            result = await userge.get_inline_bot_results(bot_, f"tweetIT {message.from_user.id} -done")
-            break
-        except:
-            current_time = time.time()
-            if current_time - start_time > 25:
-                return await message.edit("`Timeout.`", del_in=3)
+        await userge.send_message(bot_, json_)
+    try:
+        async with userge.conversation(bot_, timeout=20) as conv:
+            response = await conv.get_response(mark_read=True, filters=(filters.bot))
+    except TimeoutError:
+        return await message.edit("`Bot didn't respond...`", del_in=5)
+    resp = response.text
+    if resp != "Sticker done.":
+        return await message.edit(resp, del_in=5)
+    result = await userge.get_inline_bot_results(bot_, f"tweetIT {message.from_user.id} -done")
     await asyncio.gather(
         userge.send_inline_bot_result(
             chat_id=message.chat.id,
             query_id=result.query_id,
             result_id=result.results[0].id,
-            reply_to_message_id=reply_.message_id
+            reply_to_message_id=reply_.message_id,
         ),
-        message.delete()
+        message.delete(),
     )
 
 
@@ -78,14 +84,16 @@ async def make_tweet(message: Message):
             "-r": "add the replied message of quote message",
             "-f": "add fake text",
         },
-        "usage": "{tr}qit [reply to message]"
+        "usage": "{tr}qit [reply to message]",
     },
 )
 async def make_quote(message: Message):
     try:
         await userge.get_chat_member(-1001331162912, message.from_user.id)
-    except:
-        return await message.edit("First join **@UX_xplugin_support** and get approved by Kakashi.")
+    except BaseException:
+        return await message.edit(
+            "First join **@UX_xplugin_support** and get approved by Kakashi."
+        )
     reply_ = message.replied
     if not reply_:
         return await message.edit("`Reply to message...`", del_in=5)
@@ -108,7 +116,7 @@ async def make_quote(message: Message):
         "name": name_,
         "text": text_,
         "reply_name": reply_name,
-        "reply_text": reply_text
+        "reply_text": reply_text,
     }
     json_ = json.dumps(form_, indent=4)
     if pfp_:
@@ -116,23 +124,22 @@ async def make_quote(message: Message):
         msg_ = await userge.send_photo(bot_, down_, caption=json_)
         os.remove(down_)
     else:
-        msg_ = await userge.send_message(bot_, json_)
-    await asyncio.sleep(5)
-    start_time = time.time()
-    while True:
-        try:
-            result = await userge.get_inline_bot_results(bot_, f"quoteIT {message.from_user.id} -done")
-            break
-        except:
-            current_time = time.time()
-            if current_time - start_time > 15:
-                return await message.edit("`Timeout.`", del_in=3)
+        await userge.send_message(bot_, json_)
+    try:
+        async with userge.conversation(bot_, timeout=20) as conv:
+            response = await conv.get_response(mark_read=True, filters=(filters.bot))
+    except TimeoutError:
+        return await message.edit("`Bot didn't respond...`", del_in=5)
+    resp = response.text
+    if resp != "Sticker done.":
+        return await message.edit(resp, del_in=5)
+    result = await userge.get_inline_bot_results(bot_, f"quoteIT {message.from_user.id} -done")
     await asyncio.gather(
         userge.send_inline_bot_result(
             chat_id=message.chat.id,
             query_id=result.query_id,
             result_id=result.results[0].id,
-            reply_to_message_id=reply_.message_id
+            reply_to_message_id=reply_.message_id,
         ),
-        message.delete()
+        message.delete(),
     )
